Allowed for reading and writing to work
I made some dumb mistakes when identifying characteristics for reading and writing so I fixed those and then made new functions for reading and writing as I found that reading with a characteristic buffer is really slow actually so this now allows for the user to read and or write directly to the characteristic if they want too. This also allows the user to not have to make a characteristic or packet buffer in order to read or write.